Reorganize .graphql files
Ref DDFHER-197
