+++ web/modules/custom/bnf/bnf_server/graphql/bnf_server.base.graphqls
@@ -1,7 +1,3 @@
-extend type Mutation {
-  import(uuid: String!, callbackUrl: String!): ImportResponse!
-}
-
 type ImportResponse {
   status: ImportStatus!
   message: String!
