+++ web/modules/custom/bnf/bnf_server/graphql/bnf_server.extension.graphqls
@@ -1,0 +1,3 @@
+extend type Mutation {
+  import(uuid: String!, callbackUrl: String!): ImportResponse!
+}
