--- web/modules/custom/bnf/bnf_server/graphql/import_extension.base.graphqls
@@ -1,1 +1,0 @@
-# Base file, necessary for GraphQL module to load our extension.
